Refactor Header component
