+++ src/pages/live.js
@@ -4,9 +4,9 @@ import SEO from '../components/SEO';
 import TextSkeleton from '../components/carbon/TextSkeleton';
 import DropdownSkeleton from '../components/carbon/DropdownSkeleton';
 import Dropdown from '../components/carbon/Dropdown';
-import Header from '../components/carbon/Header';
 import SideNav from '../components/carbon/SideNav';
 import GMap from '../components/Map/GMap';
+import Header from '../components/Navbar/Header';
 import MobileNavigation from '../components/Navbar/MobileNavigation';
 import { VesselStatus } from '../components/InfoPanel';
 import configuration from '../common/api/configuration';
